MULE-16401: Fix DbFunctions internal classes exposure (#153)
